requirements.txt
google-generativeai==0.4.1  
tavily-python==0.3.1        
python-dotenv==1.0.1        
langchain==0.1.16          
langgraph==0.0.20         
faiss-cpu==1.7.4            
requests==2.31.0            
